+++ src/lib.rs
@@ -1,6 +1,7 @@
 use secp256k1::{PublicKey, SecretKey};
 use std::str::FromStr;
 
+pub mod bech32;
 pub mod events;
 pub mod keys;
 pub mod nips;
@@ -37,7 +38,14 @@ impl FromStr for Identity {
     /// assert!(identity.is_err());
     /// ```
     fn from_str(secret_key: &str) -> Result<Self, Self::Err> {
-        let secret_key = keys::secret_key_from_str(secret_key)?;
+        let secret_key = keys::secret_key_from_str(&{
+            if secret_key.starts_with("nsec") {
+                crate::bech32::from_hb_to_hex(crate::bech32::ToBech32Kind::SecretKey, secret_key)
+                    .unwrap()
+            } else {
+                secret_key.to_string()
+            }
+        })?;
         let public_key = keys::get_public_key_from_secret(&secret_key);
         let address = keys::get_str_keys_from_secret(&secret_key).1;